revolve issue where sleep max duration was limited (71 minutes !!) by uint32 type

+++ src/MQTTManager.cpp
@@ -297,7 +297,7 @@ void onMqttMessage(const char *topic, const uint8_t *payload, uint16_t length)
         if (doc.containsKey("sleep"))
         {
             DisplayManager.setPower(false);
-            PowerManager.sleep(doc["sleep"].as<uint32_t>());
+            PowerManager.sleep(doc["sleep"].as<uint64_t>());
         }
 
         delete[] payloadCopy;
